requirements.txt
qrcode==7.3.1
Cython==0.29.33
fire==0.4.0
rich==10.13.0
django-otp==1.1.3
aiofiles==0.8.0
huaweicloudsdkiam==3.0.94
aliyun-python-sdk-ram==3.2.0
aliyun-python-sdk-alidns==2.6.29
aliyun-python-sdk-cdn==3.6.4
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-dcdn==2.0.0
anyio==3.6.1
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.2
boto3==1.24.17
botocore==1.27.17
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.1.3
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.13.0
dnspython==2.2.1
dwebsocket==0.5.12
eventlet==0.33.1
executing==0.8.3
fastapi==0.78.0
filelock==3.7.1
future==0.18.2
gevent==21.12.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
huaweicloudsdkcdn==3.0.92
huaweicloudsdkcore==3.0.92
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.4
incremental==21.3.0
ipython==8.1.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
logger==1.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.25
Naked==0.1.31
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.2
prompt-toolkit==3.0.28
protobuf==4.21.1
ptyprocess==0.7.0
pure-eval==0.2.2
py3dns==3.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pydantic==1.9.1
Pygments==2.11.2
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
python-whois==0.7.3
PyYAML==6.0
requests==2.27.1
requests-futures==1.0.0
requests-toolbelt==0.9.1
resolvelib==0.5.4
s3transfer==0.6.0
service-identity==21.1.0
shellescape==3.8.1
simplejson==3.17.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
stack-data==0.2.0
starlette==0.19.1
tencentcloud-sdk-python==3.0.439
traitlets==5.1.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
uvloop==0.16.0
virtualenv==20.15.1
watchgod==0.8.2
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
ydsdk==0.2.3
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
huaweicloudsdkwaf==3.1.5
huaweicloudsdkdns